PWGJE: Reduce the storage size
