+++ PWGJE/Tasks/CMakeLists.txt
@@ -160,8 +160,8 @@ if(FastJet_FOUND)
                         SOURCES triggerCorrelations.cxx
                         PUBLIC_LINK_LIBRARIES O2::Framework O2::EMCALBase O2::EMCALCalib O2Physics::PWGJECore O2Physics::AnalysisCore
                         COMPONENT_NAME Analysis)
-    o2physics_add_dpl_workflow(jet-charged-trigger-qa
-                        SOURCES ChJetTriggerQATask.cxx
+    o2physics_add_dpl_workflow(jet-trigger-charged-qa
+                        SOURCES jetTriggerChargedQA.cxx
                         PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::PWGJECore O2Physics::AnalysisCore
                         COMPONENT_NAME Analysis)
     o2physics_add_dpl_workflow(jet-full-trigger-qa
