+++ PWGJE/Tasks/recoilJets.cxx
@@ -557,7 +557,7 @@ struct RecoilJets {
   template <typename Collision>
   bool skipMBGapEvent(const Collision& coll)
   {
-    return (skipMBGapEvents && coll.subGeneratorId()) == jetderiveddatautilities::JCollisionSubGeneratorId::mbGap;
+    return skipMBGapEvents && coll.subGeneratorId() == jetderiveddatautilities::JCollisionSubGeneratorId::mbGap;
   }
 
   template <typename Track>
